pyproject.toml
[tool.poetry]
name = "streamlit-sales-suite"
version = "0.1.0"
description = "An all-encompassing web platform for data analysis and visualization, powered by Streamlit, complemented by an automated item classifier that streamlines the process of label assignment."
authors = ["danibcorr <danibcorr@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.37.1"
deep-translator = "^1.11.4"
pandas = "^2.2.2"
tensorflow = {version = "2.16.2", extras = ["and-cuda"]}
pillow = "^10.4.0"
plotly = "^5.23.0"
opencv-python = "^4.10.0.84"
scikit-image = "^0.24.0"
tqdm = "^4.66.5"
bing-image-downloader = "^1.1.2"
mlflow = "^2.15.1"
loguru = "^0.7.2"
keras-cv = "^0.9.0"
pynvml = "^11.5.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
